Fix for /announce
As the previous version did not allow to have the text first, we had to
move it to the end. During so the optional paramters now became
required. Also the time has been added to be given. The command
paramters have also been added in /help. Additonal to this, the typo for
the alias has been fixed.

Notice: There is a restriction for styles, we can not use style 2 as it
is bugged.